requirements.txt
playwright==1.44.0
httpx==0.27.0
beautifulsoup4==4.12.3
googlemaps==4.10.0
gspread==6.1.2
google-auth==2.29.0
anthropic==0.28.0
openai==1.30.1
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
click==8.1.7
